refactor:modify query to include truncate and move variable to top
